+++ plugin/src/test/kotlin/ru/cian/huawei/publish/utils/ConfigProviderTest.kt
@@ -1,7 +1,9 @@
 package ru.cian.huawei.publish.utils
 
+import assertk.assertFailure
 import assertk.assertThat
 import assertk.assertions.isEqualTo
+import assertk.assertions.messageContains
 import assertk.tableOf
 import io.mockk.every
 import io.mockk.mockk
@@ -21,13 +23,13 @@ import ru.cian.huawei.publish.HuaweiPublishConfig
 import ru.cian.huawei.publish.HuaweiPublishExtensionConfig
 import ru.cian.huawei.publish.ReleasePhaseConfig
 import ru.cian.huawei.publish.ReleasePhaseExtension
-import ru.cian.huawei.publish.models.Credential
 import java.io.File
 import ru.cian.huawei.publish.ReleaseNote
 import ru.cian.huawei.publish.ReleaseNotesConfig
 import ru.cian.huawei.publish.ReleaseNotesExtension
 import ru.cian.huawei.publish.ReleaseNotesDescriptionsConfig
 
+private const val DEFAULT_PUBLISH_SOCKET_TIMEOUT_IN_SECONDS = 60L
 private const val DEFAULT_PUBLISH_TIMEOUT_MS = 10 * 60 * 1000L
 private const val DEFAULT_PUBLISH_PERIOD_MS = 15 * 1000L
 private const val BUILD_DIRECTORY_PATH = "./build"
@@ -49,6 +51,7 @@ private const val CREDENTIALS_SECOND_JSON = "{\"client_id\": \"no_id\", \"client
 private const val APP_BASIC_INFO_FILE_PATH = "$BUILD_DIRECTORY_PATH/app_info.json"
 private const val APP_BASIC_INFO_FILE_SECOND_PATH = "$BUILD_DIRECTORY_PATH/app_info_second.json"
 
+@Suppress("LongMethod")
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 internal class ConfigProviderTest {
 
@@ -128,17 +131,19 @@ internal class ConfigProviderTest {
             releaseNotesFileProvider = releaseNotesFileProvider,
         )
 
-        assertThat { configProvider.getConfig() }.hasException(IllegalArgumentException::class)
+        assertFailure { configProvider.getConfig() }
+            .messageContains("has wrong file extension that doesn't match with announced buildFormat(APK) plugin extension param")
     }
 
     @Test
-    fun `correct config for default params`() = mockkObject(CredentialHelper) {
+    fun `correct config with overriding credentials params`() = mockkObject(CredentialHelper) {
 
         val expectedConfig = HuaweiPublishConfig(
             credentials = Credentials("id", "secret"),
             deployType = DeployType.PUBLISH,
             artifactFormat = BuildFormat.APK,
             artifactFile = File(ARTIFACT_APK_FILE_PATH),
+            publishSocketTimeoutInSeconds = DEFAULT_PUBLISH_SOCKET_TIMEOUT_IN_SECONDS,
             publishTimeoutMs = DEFAULT_PUBLISH_TIMEOUT_MS,
             publishPeriodMs = DEFAULT_PUBLISH_PERIOD_MS,
             releaseTime = null,
@@ -147,11 +152,7 @@ internal class ConfigProviderTest {
             appBasicInfoFile = null
         )
 
-        every {
-            CredentialHelper.getCredentials(match { it.absolutePath == CREDENTIALS_FILE_PATH })
-        } returns Credential(clientId = "id", clientSecret = "secret")
-
-        tableOf("expectedValue", "actualValue")
+        tableOf("expectedValue", "configProvider")
             .row(
                 expectedConfig,
                 ConfigProvider(
@@ -162,18 +163,118 @@ internal class ConfigProviderTest {
                 )
             )
             .row(
-                expectedConfig,
+                expectedConfig.copy(
+                    credentials = Credentials("no_id", "no_secret")
+                ),
+                ConfigProvider(
+                    extension = extensionConfigInstance().apply {
+                        credentialsPath = CREDENTIALS_FILE_SECOND_PATH
+                    },
+                    cli = emptyCliConfig,
+                    buildFileProvider = buildFileProvider,
+                    releaseNotesFileProvider = releaseNotesFileProvider,
+                )
+            )
+            .row(
+                expectedConfig.copy(
+                    credentials = Credentials("no_id", "no_secret")
+                ),
                 ConfigProvider(
                     extension = extensionConfigInstance(),
                     cli = HuaweiPublishCliParam(
-                        credentialsPath = CREDENTIALS_FILE_PATH
+                        credentialsPath = CREDENTIALS_FILE_SECOND_PATH
                     ),
                     buildFileProvider = buildFileProvider,
                     releaseNotesFileProvider = releaseNotesFileProvider,
                 )
             )
-            .forAll { expectedValue, actualValue ->
-                assertThat(actualValue.getConfig()).isEqualTo(expectedValue)
+            .row(
+                expectedConfig.copy(
+                    credentials = Credentials("no_id", "no_secret")
+                ),
+                ConfigProvider(
+                    extension = extensionConfigInstance().apply {
+                        credentialsPath = CREDENTIALS_FILE_PATH
+                    },
+                    cli = HuaweiPublishCliParam(
+                        credentialsPath = CREDENTIALS_FILE_SECOND_PATH
+                    ),
+                    buildFileProvider = buildFileProvider,
+                    releaseNotesFileProvider = releaseNotesFileProvider,
+                )
+            )
+            .row(
+                expectedConfig.copy(
+                    credentials = Credentials("id_base64", "secret_base64")
+                ),
+                ConfigProvider(
+                    extension = extensionConfigInstance().apply {
+                        credentials = "ewogICJjbGllbnRfaWQiOiAiaWRfYmFzZTY0IiwKICAiY2xpZW50X3NlY3JldCI6ICJzZWNyZXRfYmFzZTY0Igp9"
+                    },
+                    cli = emptyCliConfig,
+                    buildFileProvider = buildFileProvider,
+                    releaseNotesFileProvider = releaseNotesFileProvider,
+                )
+            )
+            .row(
+                expectedConfig.copy(
+                    credentials = Credentials("id_base64", "secret_base64")
+                ),
+                ConfigProvider(
+                    extension = extensionConfigInstance(),
+                    cli = HuaweiPublishCliParam(
+                        credentials = "ewogICJjbGllbnRfaWQiOiAiaWRfYmFzZTY0IiwKICAiY2xpZW50X3NlY3JldCI6ICJzZWNyZXRfYmFzZTY0Igp9"
+                    ),
+                    buildFileProvider = buildFileProvider,
+                    releaseNotesFileProvider = releaseNotesFileProvider,
+                )
+            )
+            .row(
+                expectedConfig.copy(
+                    credentials = Credentials("id_base64", "secret_base64")
+                ),
+                ConfigProvider(
+                    extension = extensionConfigInstance().apply {
+                        credentials = "must_not_be_used"
+                    },
+                    cli = HuaweiPublishCliParam(
+                        credentials = "ewogICJjbGllbnRfaWQiOiAiaWRfYmFzZTY0IiwKICAiY2xpZW50X3NlY3JldCI6ICJzZWNyZXRfYmFzZTY0Igp9"
+                    ),
+                    buildFileProvider = buildFileProvider,
+                    releaseNotesFileProvider = releaseNotesFileProvider,
+                )
+            )
+            .row(
+                expectedConfig.copy(
+                    credentials = Credentials("id_base64", "secret_base64")
+                ),
+                ConfigProvider(
+                    extension = extensionConfigInstance().apply {
+                        credentials = "ewogICJjbGllbnRfaWQiOiAiaWRfYmFzZTY0IiwKICAiY2xpZW50X3NlY3JldCI6ICJzZWNyZXRfYmFzZTY0Igp9"
+                    },
+                    cli = HuaweiPublishCliParam(
+                        credentialsPath = CREDENTIALS_FILE_SECOND_PATH
+                    ),
+                    buildFileProvider = buildFileProvider,
+                    releaseNotesFileProvider = releaseNotesFileProvider,
+                )
+            )
+            .row(
+                expectedConfig.copy(
+                    credentials = Credentials("no_id", "no_secret")
+                ),
+                ConfigProvider(
+                    extension = extensionConfigInstance().apply {
+                        credentialsPath = CREDENTIALS_FILE_SECOND_PATH
+                    },
+                    cli = emptyCliConfig,
+                    buildFileProvider = buildFileProvider,
+                    releaseNotesFileProvider = releaseNotesFileProvider,
+                )
+            )
+            .forAll { expectedValue, configProvider ->
+                val actualValue = configProvider.getConfig()
+                assertThat(actualValue).isEqualTo(expectedValue)
             }
     }
 
@@ -181,10 +282,11 @@ internal class ConfigProviderTest {
     fun `correct config with overriding common values at cli params`() {
 
         val expectedConfig = HuaweiPublishConfig(
-            credentials = Credentials("id123", "secret123"),
+            credentials = Credentials("id_base64", "secret_base64"),
             deployType = DeployType.DRAFT,
             artifactFormat = BuildFormat.AAB,
             artifactFile = File(ARTIFACT_AAB_FILE_SECOND_PATH),
+            publishSocketTimeoutInSeconds = 3003L,
             publishTimeoutMs = 1001L,
             publishPeriodMs = 2002L,
             releaseTime = "2019-10-18T21:00:00+0300",
@@ -198,7 +300,7 @@ internal class ConfigProviderTest {
         )
 
         val inputExtensionConfig = extensionConfigInstance().apply {
-            credentialsPath = CREDENTIALS_FILE_PATH
+            credentialsPath = CREDENTIALS_FILE_SECOND_PATH
             deployType = DeployType.PUBLISH
             publishTimeoutMs = 3003
             publishPeriodMs = 4004
@@ -213,12 +315,12 @@ internal class ConfigProviderTest {
             appBasicInfo = APP_BASIC_INFO_FILE_PATH
         }
         val inputCliConfig = HuaweiPublishCliParam(
+            credentials = "ewogICJjbGllbnRfaWQiOiAiaWRfYmFzZTY0IiwKICAiY2xpZW50X3NlY3JldCI6ICJzZWNyZXRfYmFzZTY0Igp9",
             deployType = DeployType.DRAFT,
+            publishSocketTimeoutInSeconds = "3003",
             publishTimeoutMs = "1001",
             publishPeriodMs = "2002",
             credentialsPath = CREDENTIALS_FILE_SECOND_PATH,
-            clientId = "id123",
-            clientSecret = "secret123",
             buildFormat = BuildFormat.AAB,
             buildFile = ARTIFACT_AAB_FILE_SECOND_PATH,
             releaseTime = "2019-10-18T21:00:00+0300",
@@ -248,6 +350,7 @@ internal class ConfigProviderTest {
             deployType = DeployType.PUBLISH,
             artifactFormat = BuildFormat.APK,
             artifactFile = File(ARTIFACT_APK_FILE_PATH),
+            publishSocketTimeoutInSeconds = DEFAULT_PUBLISH_SOCKET_TIMEOUT_IN_SECONDS,
             publishTimeoutMs = DEFAULT_PUBLISH_TIMEOUT_MS,
             publishPeriodMs = DEFAULT_PUBLISH_PERIOD_MS,
             releaseTime = null,
@@ -319,7 +422,6 @@ internal class ConfigProviderTest {
             }
     }
 
-    @Suppress("LongMethod")
     @Test
     fun `correct config with overriding release notes`() {
         val expectedConfig = HuaweiPublishConfig(
@@ -327,6 +429,7 @@ internal class ConfigProviderTest {
             deployType = DeployType.PUBLISH,
             artifactFormat = BuildFormat.APK,
             artifactFile = File(ARTIFACT_APK_FILE_PATH),
+            publishSocketTimeoutInSeconds = DEFAULT_PUBLISH_SOCKET_TIMEOUT_IN_SECONDS,
             publishTimeoutMs = DEFAULT_PUBLISH_TIMEOUT_MS,
             publishPeriodMs = DEFAULT_PUBLISH_PERIOD_MS,
             releaseTime = null,